Add anvil-changing player to anvil update event
If a mod alters the output of an anvil, then much vanilla behavior will be skipped inside ContainerRepair#updateRepairOutput(). Some of that anvil behavior depends on the player (i.e. creative mode capability). If the mod wants to preserve (imitate) that behavior, the mod needs data not currently available to it. It needs at least a creative-mode boolean, but since some future mod could design further anvil behavior based on other player data, I propose to pass the whole player reference (ContainerRepair's this.thePlayer).

This change to the AnvilUpdateEvent is the first of (I think) three parts. The other two will be ForgeHooks#onAnvilChange where the revised event is instantiated, and vanilla ContainerRepair's call of the hook.